Simple spring security authorization is added

+++ src/main/java/testgroup/philmography/dao/FilmDao.java
@@ -9,5 +9,5 @@ public interface FilmDao {
 	void add(Film film);
 	void delete(Film film);
 	void update(Film film);
-	Film getById(int id);
+	Film getById(Long id);
 }
